removed old UserRepository and added new one with JPA

+++ src/main/java/com/example/virtualwallet/models/dtos/pageable/UserPageOutput.java
@@ -10,7 +10,7 @@ public class UserPageOutput {
 
     int numberOfPages;
 
-    int totalResults;
+    long totalResults;
 
     List<UserOutput> content;
 }
